requirements.txt

aiogram~=2.11.2
opencv-python~=4.5.1.48
numpy~=1.20.1
Pillow~=8.1.1
pymongo~=3.11.4
APScheduler~=3.6.3
SQLAlchemy~=1.3.23